# Closes #1904: Add Chapel-native encoding/decoding functionality