--- a/dep/checkIconv.chpl
+++ b/dep/checkIconv.chpl
@@ -1,0 +1,5 @@
+use iconv;
+
+proc main() {
+  writeln("Found libiconv version: ", _libiconv_version);
+}