--- a/src/Codecs.chpl
+++ b/src/Codecs.chpl
@@ -108,4 +108,3 @@ module Codecs {
       }
     }
 }
-