--- a/src/EncodingMsg.chpl
+++ b/src/EncodingMsg.chpl
@@ -47,14 +47,9 @@ module EncodingMsg {
       var encodeOffsets = (+ scan encodeLengths);
       encodeOffsets -= encodeLengths;
       
-      var finalValues = makeDistArray(+ reduce encodeLengths, uint(8));
-      forall (i, off, len) in zip(0..#stringsObj.size, offs, lengths) {
-        var slice = new lowLevelLocalizingSlice(origVals, off..#len);
-        var encodedStr = encodeStr(slice.ptr: c_ptr(uint(8)), len, encodeLengths[i], toEncoding, fromEncoding);
-        finalValues[encodeOffsets[i]..#encodeLengths[i]] = encodedStr;
-      }
+      var encodedValues = encodeSegments(offs, origVals, encodeLengths, toEncoding, fromEncoding);
 
-      return (encodeOffsets, finalValues);
+      return (encodeOffsets, encodedValues);
     }
 
     proc getBufLengths(segments: [?D] int, values: [?vD] ?t, toEncoding: string, fromEncoding: string) throws {
@@ -82,6 +77,48 @@ module EncodingMsg {
             forall (start, len, i) in zip(mySegs, myLens, mySegInds) with (var agg = newDstAggregator(int)) {
               var slice = new lowLevelLocalizingSlice(values, start..#len);
               agg.copy(res[i], getBufLength(slice.ptr: c_ptr(uint(8)), len, toEncoding, fromEncoding));
+            }
+          } catch {
+            throw new owned ErrorWithContext("Error",
+                                             getLineNumber(),
+                                             getRoutineName(),
+                                             getModuleName(),
+                                             "IllegalArgumentError");
+          }
+        }
+      }
+      return res;
+    }
+
+    proc encodeSegments(segments: [?D] int, values: [?vD] uint(8), encodeLengths: [D] int, toEncoding: string, fromEncoding: string) throws {
+      var res = makeDistArray(+ reduce encodeLengths, uint(8));
+      if (D.size == 0) {
+        return res;
+      }
+
+      const (startSegInds, numSegs, lengths) = computeSegmentOwnership(segments, vD);
+    
+      // Start task parallelism
+      coforall loc in Locales {
+        on loc {
+          const myFirstSegIdx = startSegInds[loc.id];
+          const myNumSegs = max(0, numSegs[loc.id]);
+          const mySegInds = {myFirstSegIdx..#myNumSegs};
+          // Segment offsets whose bytes are owned by loc
+          // Lengths of segments whose bytes are owned by loc
+          var mySegs, myLens: [mySegInds] int;
+          forall i in mySegInds with (var agg = new SrcAggregator(int)) {
+            agg.copy(mySegs[i], segments[i]);
+            agg.copy(myLens[i], lengths[i]);
+          }
+          try {
+            forall (start, len, i, encodeLen) in zip(mySegs, myLens, mySegInds, encodeLengths) {
+              var slice = new lowLevelLocalizingSlice(values, start..#len);
+              var encodedStr = encodeStr(slice.ptr: c_ptr(uint(8)), len, encodeLen, toEncoding, fromEncoding);
+              var agg = newDstAggregator(uint(8));
+              for j in start..#len {
+                agg.copy(res[j], encodedStr[j-start]);
+              }
             }
           } catch {
             throw new owned ErrorWithContext("Error",