This PR adds support for the iconv and idn2 libraries, which enable Chapel-native support for encoding and decoding strings from IDNA and a number of other encodings. This results in a significant performance improvement (~1000x) over the current implementation that calls out to Python to handle the encoding.

The additional idn2 and iconv libraries are now required Arkouda dependencies and can be installed with `make install-iconv` and `make install-idn2`.

Performance numbers collected on a 16-node-cs-hdr using 1,000,000 strings using strings normally distributed between lengths 1 to 16:

Locales | Original | Native Chapel | Optimized Native
------- | -------- | ------------- | ----------------
1       |  ~5,000s |            6s |           0.12s
2       | ~10,000s |          110s |           0.21s
4       |  --      |            -- |           0.24s
8       |  --      |            -- |           0.26s

Where "Optimized Native" is the current version of this PR and "Native Chapel" was the performance of inserting Chapel-native calls into the code from master prior to this PR.

closes #1904