@bmcdonald3 2022-11-14
@Ethan-DeBandi99 that sounds good to me WRT defaulting to UTF-8 for encode fromEncoding and then forcing input for toEncoding.